lab 1 fixed 2.0
