+++ tellez.aron/P1/functions.hpp
@@ -1,0 +1,8 @@
+#ifndef GET_SEQUENCE_H
+#define GET_SEQUENCE_H
+
+int getNextValue();
+int updateSequenceCounts(int max_seq_count, int current_seq_count);
+int getSequence();
+
+#endif
