--- tellez.aron/P1/getSequence.hpp
@@ -1,6 +1,0 @@
-#ifndef GET_SEQUENCE_H
-#define GET_SEQUENCE_H
-
-int getSequence();
-
-#endif
